added Battery Monitor Simulator, refactored to battery monitor to a base class
